broken data

+++ week3/d01/Assignment_3_1_2/includes/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-int random_num(int a, int b);
+double random_num(double a, double b);
 void clear_cin();
 
 #endif
